requirements.txt
asgiref==3.5.0
crispy-bootstrap5==0.6
Django==3.2
django-autoslug==1.9.8
django-ckeditor==6.3.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-js-asset==2.0.0
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
